refactor(seed): own committed seed explanations
